attempt at requirejs but toBase64 function intermitten failures
