+++ app/main.js
@@ -1,0 +1,6 @@
+require.config({
+    baseUrl: 'js/lib',
+    paths: {
+        jquery: 'jquery'
+    }
+});
