+++ index.html
@@ -15,6 +15,8 @@
   <link rel="stylesheet" type="text/css" href="css/animate.css">
   <link rel="stylesheet" type="text/css" href="css/styles.css" />
 
+  <script data-main="js/app" src="js/require.js"></script>
+  <!--
   <script type="text/javascript" language="javascript" src="js/jquery.js"></script>
   <script type="text/javascript" language="javascript" src="js/jquery.dataTables.min.js"></script>
   <script type="text/javascript" language="javascript" src="js/jquery.cookie.js"></script>
@@ -28,6 +30,7 @@
   <script type="text/javascript" language="javascript" src="js/cryptojs_v3.1.2.min.js"></script>
   <script type="text/javascript" language="javascript" src="js/xapiwrapper.min.js"></script>
   <script type="text/javascript" language="javascript" src="js/scripts.js" class="init"></script>
+  -->
 </head>
 
 <body>
