+++ js/lib/collapse.js
