+++ js/lib/jquery.cookie.js
