+++ js/lib/transition.js
